[simi-review-tool] feat: Optimize review tool
