+++ docs/Japanese.md
@@ -1,0 +1,3 @@
+## February 2025
+言葉 (ことば,kotoba)  
+読む (よみ,yomu)
