+++ docs/Questions.md
@@ -1,0 +1,9 @@
+# Table of Contents
+[Back to Main Project README](../README.md)  
+- [Java](#java)
+  - [1. How does `AbstractQueuedSynchronizer` work?](#1-how-does-abstractqueuedsynchronizer-work) 
+# Java
+##  1. How does `AbstractQueuedSynchronizer` function internally?
+### References
+* **Java.pdf / Java / java.base / java.util / concurrent / locks / AbstractQueuedSynchronizer**     
+
