--- docs/review-logs.txt
@@ -1,4 +1,0 @@
-2025-02-18 23:11:14 - Algorithm Problems / Precomputation / 21. Range Product Queries of Powers
-2025-02-18 23:14:47 - Algorithm Problems / Traversal / 24. Find the Number of Ways to Place People I
-2025-02-18 23:19:19 - Algorithm Problems / String / 23. License Key Formatting
-2025-02-18 23:19:22 - English / Near-synonym / SSS / same, similar, resemble
